Support overriding programs in conformance tests (#18787)
This change adds the ability to specific that one or more conformance
tests in a suite should use hardcoded or pre-generated programs, rather
than relying on a language host's `GenerateProject` functionality. This
is useful in cases where e.g. a language's core SDK supports
functionality which we'd like to test, but its code generation features
are not fleshed out enough to produce programs capable of exercising
those features -- Java is a good example of this. With this feature,
such tests can now be made to pass by hand-writing the relevant programs
and supplying these paths as overrides. Aside from not generating a
program, all other functionality -- snapshot checking, assertions, etc.
is the same. This is verified with a metatest that overrides the program
for the `l1-empty` with a mocked language host.

Fixes #18235