Extend `Handshake` responses to allow more configuration
When we initially introduced the `Handshake` gRPC method for negotiating
communication with providers, we anticipated that any provider
implementing `Handshake` would implicitly support all the features we
had until that point supported flagging in `Configure`. However, this
turns out not to be the case -- the bridge, for instance, supports the
very new autonaming but still doesn't support secrets (or at least,
indicate that it does). This PR thus gives providers the ability to
return a bit more fine-grained configuration in the `Handshake`
response, hopefully smoothing the path to introduction/broadly making
`Handshake` more useful. Note that we leave `SupportsPreview` implicit
-- this is a bit of a hand-wavy decision since the bridge supports it
and the bridge underpins the majority of our providers.