Test state delete directly via cobra (#18788)
Now that @iwahbe has made our use of cobra more standard we can just
test directly against the cobra `Command` rather than having to mock it
via a separate struct. This does that work for the state delete command,
just cause it looked like an easy one to cleanup.

With this we've no longer got a separate struct, everything is contained
within the `newStateDeleteCommand` which returns a `*cobra.Command`
which is directly testable.

This also ensures we're testing args parsing and expected cobra logic
(like context setting) as well.