Make protect optional
This is so we can correctly inherit the setting from our parent, but still
allow child resources to override it. If a parent has protect set to true, and
a child explictly sets it to false then protect _should not_ be set on the
child.

Currently because we only have true/false we treat false the same as not set,
and so a child resource can't override a parents true setting.