Update from exp/maps to maps in the conformance tests
These changes require Go 1.23 which should be fine to upgrade
conformance tests to. We could make similar changes across pkg and sdk
but might be a bit aggressive forcing them to 1.23 this soon (1.24 only
just came out, so even though 1.22 technicaly isn't in our support list
anymore we probably don't want to be too quick to force 1.23 on people).