Add the option to run the program on destroy (#18751)
Fixes https://github.com/pulumi/pulumi/issues/18763.

This doesn't actually expose this to users, just plumbing in the
internals. This adds a new DestroyV2 operation which behaves like a
normal update operation except it puts the step generator into a new
"destroy" mode.

In destroy mode the step generator behaves mostly like normal for
component resources and provider resources, but all custom resources are
skipped over. Either becoming same steps if they have old state, or
skipped creates if they don't. All resources (custom, component, and
providers) are all tagged for deletion for at the end of step
generation.
That is normally ScheduleDeletes that runs at the end of the program
just deletes things the step generator hasn't seen yet, but in destroy
mode it deletes things the step generator hasn't seen yet _and_ all the
new things it has seen.

Even though we're running the program we're not actually going to
create/update any real resources, only providers and components.
Components are just so that they resolve in the program as much as
possible to allow it to keep running. Providers so that we can update
credentials or provider versions. That being the big difference between
running destroy with vs without a program.

All of this still respects `--targets`.