print all diagnostic errors when getting the schema failed (#18817)
When getting the schema fails and returns a diagnostics error we
currently only print the first error, and say "x more diagnostics". This
happens e.g. when the schema fails to validate against the pulumi.json
schema. It's often helpful to see the entire list of errors instead, as
that helps seeing the whole picture. Print the whole list instead.