Make `whoami` just a cobra command (#18845)
This removes the `whoAmICmd` to replace it with just the cobra wrapper. 

There's three things worth calling out here.
1. We can still mock out stdout by using `cmd.OutOrStdout()` in the run
function, and setting it explicitly in the tests. Otherwise cobra
defaults it to `os.Stdout`. This will be the general way of mocking
stdout in commands going forward.
2. We inject the `workspace.Context` and `backend.LoginManager` to the
new function so we can mock them in tests. They were hardcoded in the
run function to the singleton instances.
3. Because we're injecting those two types we can just use the global
`CurrentBackground` function. If we do this everywhere we should be able
to remove the global `BackendInstance` variable from that function
eventually.