Optimize crypters via batch encrypting & cache unification (#18743)
Improves performance for end-users of secrets in state by:
1. Making a round-trip to the service for each secret as it's found
during serialization is slow. Therefore, we want to find all the secrets
we're going to serialize, fetch them in a single batch request to the
service.
2. If we've paid the cost to decrypt a secret and have not modified it,
we don't need to go back to the service to re-encrypt it. If no secrets
get modified, so we can completely skip the remote call to do encryption
and use the original cached ciphertext for the same instance of the
secret (but not other secrets with the same value).

We introduce three new components to unify and formalise how crypters
can be optimised for batch operations.

Firstly, we introduce a `SecretCache` which models when previous encrypt
or decrypt operations are permitted to be re-used. This can be accessed
by either `*secret` + `plaintext` giving a `ciphertext` if the plaintext
still matches the original cached plaintext for the specific secret
instance, or by `ciphertext`, returning the associated `plaintext`. This
class is fully thread-safe and can be safely persisted and shared
between multiple crypters.

Secondly, we create a `BatchEncrypter` and `BatchDecrypter` interface
which extends the underlying `Encrypter` and `Decrypter` interfaces,
respectively. These are created by calling a `BeginBatch[En|De]crypt`
method which returns an instance of the batch crypter class and a
`CompleteCrypterBatch` function. Each of these has an `Enqueue` method
to register an encrypt or decrypt operation for a specific secret
instance (pointer). The actual crypter operation is then performed via a
batch operation once the completion function has been called.

Instances of the `BatchEncrypter` and `BatchDecrypter` classes created
from the `CachingSecretsManager` are _caching_ batch crypters, which
also utilise a shared `SecretCache` maintained within the
`CachingSecretsManager` to avoid having to re-encrypt or re-decrypt all
secret values for operations which haven't made any changes to secrets.

- Fixes https://github.com/pulumi/home/issues/3896
- Fixes https://github.com/pulumi/home/issues/3898
- Fixes https://github.com/pulumi/pulumi/issues/15538
- Fixes https://github.com/pulumi/pulumi/issues/15293
- Fixes https://github.com/pulumi/pulumi/issues/17660
- Replaces https://github.com/pulumi/pulumi/pull/18576

## Thread Safety

The `SecretCache` uses `sync.Map` instances to ensure concurrent usage
is safe.

The queue within the batch crypters uses buffered channels to build up
the batch ready to send. Once the channel is full, we handle the failed
channel send and trigger a batch send. This will take a sending mutex
lock then send one batch (up to the maximum configured batch size) from
the queue (this will allow other threads to start writing to the channel
again in the background), complete the batch encryption, write the
results back to the target objects the unblock the mutex to allow
another thread to send if required. Once all secrets have been enqueued,
the creator of the batch must call complete exactly once, which will
block any subsequent secrets being enqueued or complete being called
again, then will process the final batch of secrets.

## Benchmark

Scenario: export 100 secret outputs from a python program.
Environment: Test run on laptop in UK against [personal review
stack](https://api-danielrbradley.review-stacks.pulumi-dev.io/).
The "Unified cache + batch" test was run on a different day with
seemingly slower baseline roundtrip performance.

| Test | initial_preview | initial_update | preview_empty | update_empty
|

|--------------|-----------------|----------------|---------------|--------------|
| Without batch | 2.10s | 19.94s | 2.44s | 20.02s |
| With batch encrypt | 2.18s | 2.86s | 2.32s | 3.08s |
| Unified cache + batch | 2.83s | 3.14s | 2.57s | 2.67s |

The change when adding batch encryption only is most significant as
we're sending a request for every secret, for secret managers that
support it.

The change when also adding the unified cache can be seen by comparing
`initial_update` with `update_empty` - if we've got no new secrets, then
we can skip the whole of the encryption round trip. This will also
affect users of secret managers that don't support batching much more
significantly - such as self-hosted customers on older (or custom) http
backends.

CI gate is blocked on the service production deployment following the
rename of the capability in https://github.com/pulumi/pulumi/pull/18776

## Service performance impact

~This feature could increase the burstability of KMS requests when
serializing deployments with a lot of secrets. It doesn't increase the
overall number of encryptions, but we could end up doing more in a short
space of time. This brings encryptions into a similar pattern to the
batch decryptions which also cause bursts of decryption requests.
However, this also should be less of an issue than the batch decryptions
as we're also optimising the engine to only require service encrypt
calls for secrets which have actually changed value, as we can just
round-trip unchanged secrets. In practice, most stacks have less than 25
secrets, so this impact is likely to be minimal for the majority of
customers. Nevertheless, we should be aware to monitor our peak quota
usages.~

Update: on further review of the service implementation this is
inaccurate as we don't perform remote KMS request for this operation.

## Skipped re-encryption impact

~Unchanged secrets within resources are not re-encrypted and therefore
could remain on old key versions for longer, just like if retrieving an
older version of the stack which hasn't been updated in the same amount
of time as the resource with the secret.~

Update: We now avoid mixing cached responses with re-encryptions. The
cache is only used if no secrets have been changed. If any secret in the
stack is changed, all secrets will be re-encrypted.

> Do we ever need to force the re-encryption of all secrets within a
stack at any point in time? If so, what should the mechanism be for
skipping the cache?

Added a temporary environment variable `PULUMI_DISABLE_SECRET_CACHE`
which can be set to `true` to completely disable the unified cache and
ensure all secrets are re-encrypted, in case of any unforeseen impacts
of increasing our cache utilisation.

`pulumi stack change-secrets-provider` is not affected by this as it
constructs separate old and new secret managers which have isolated
secret caches.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>