Fix timestamp consistency between JSON output and engine event log (#18838)
Fixes https://github.com/pulumi/pulumi/issues/18837

Previous to this change `startEventLogger` and `ShowJSONEvents` did
their own separate timestamping. This could lead to small differences in
timestamps written to the event log and those written to the display.

This moves the timestamping to be done up-front in one channel that is
then fed through both outputs, so we'll always get consistent
timestamps.

Managing all the channels here is a bit awkward because all the other
display functions just want to use normal non-timestamped engine events.