Re-add provider error (#18854)
In https://github.com/pulumi/pulumi/pull/16241/ we originally added
package references via the `provider` field. We would check to see if
the field contained a "::" and if so treat it as an explicit provider
reference, else we'd treat it as a package reference. If either of these
failed to parse/lookup it would return an error.

In https://github.com/pulumi/pulumi/pull/16458/ we added a new explicit
package ref field and stopped using the `provider` field to transmit the
package reference. When we did this we removed the case from provider
parsing for package refs, but we left in the check to see if the field
contained a "::".

So before both these changes if you sent a malformed field like "hi" it
would error, but after these two changes it would just ignore the
provider field and use the default provider instead.