use full Go version in pulumi-test-language/go.mod (#18844)
When one does not have the correct Go version installed by default on
the system, the version specified in the `go.mod` file needs to include
the patch version for the Go tooling to automatically install the
toolchain. Make that so, so things don't break for people that have
older Go versions still installed on the system.