[python/components] Better error message for a missing type (#18876)
I audited the python analyzer and it looks like error messages are
already decent (on the level of TypeScript). We can definitely do better
but at least we give the context of type/property names most of the
time.

I found just one exception - so here is a tiny change towards
https://github.com/pulumi/pulumi/issues/18692