Run lifecycle tests with parallelism (#18878)
Noticed this while working on parallel diff. Most of the lifecycle tests
leave `Parallel` as zero, which eventually makes its way down to step
executor setup which sets it to 1. So none of our lifecycle tests are
really testing any parallelism in the system.

This sets `Parallel` to `runtime.NumCPU()` so that tests will start
seeing operations happen in parallel, hopefully proving out more of the
system.

N.B. This isn't the same as what we default `--parallel` to, we respect
cgroup limits there but I figure for tests this is fine and saves moving
that code to be shared by lifecycle tests.