Remove query (#18891)
First part of https://github.com/pulumi/pulumi/issues/16964. This
removes the `query` command and all the associated logic in pkg.

This does not try to remove `StreamInvoke` and all the associated logic
for that from sdk. That should be safe to do, but I'm staggering this
just in case.