Bump several Go dependencies (#18898)
We have advisories for some of these, but they are present in tests for
the most part and so are low risk. That said, this bumps them to clean
up the list and also to align more the test dependencies, in theory
reducing the amount of work we do in CI to download Go dependencies when
running pipelines.