python: import packages from <namespace>_ (#18418)
Curently when using `package add`, packages are always namespaced in the
"pulumi" namespace. For python this means they are always named
`pulumi_<name>`. However it would be better to namespace them by
publisher if that was given, so we can disambiguate them better. This PR
makes that so for Python. Subsequent PRs are going to implement the same
thing for other languages.

Part of https://github.com/pulumi/pulumi/issues/18369