Bump several NodeJS dependencies (#18899)
We have advisories for some of these, but they are present in tests and
so are low risk. That said, this bumps them to clean up the list and
also to align more the test dependencies, in theory reducing the amount
of work we do in CI to download NPM dependencies when running pipelines.