Log safely in Python in the presence of `PULUMI_ERROR_OUTPUT_STRING`
Attempting to convert an `Output` to a string is a common pitfall Pulumi
users face -- since an `Output` value may not be available yet (akin to
a `Promise` in NodeJS, for instance), it's not in general possible or
safe to serialize an `Output` to a string in all cases. To this end, the
various language SDKs override string serialization for `Output` so that
an informative message is printed, which hopefully helps the user
achieve what they set out to (e.g. via `__str__` in Python, `toString`
in Java, and so on).

For users that want their programs to throw a hard error when they
mistakenly try to stringify an `Output`, Pulumi offers the
`PULUMI_ERROR_OUTPUT_STRING` environment variable. If this is set, the
relevant implementations (`__str__`, etc.) will crash if called. This is
all well and good, as long as we (Pulumi itself) don't mistakenly invoke
these methods and break user programs.

Alas, this is precisely what happens in #18863. Here, some debug logging
statements in the Python SDK include a value that _could_ be an `Output`
(`opts.id`). As a result, if a suitable program is run with
`PULUMI_ERROR_OUTPUT_STRING` set, Pulumi will throw an error despite the
user having done nothing wrong.

The fix is much the same as adopted in #18016 for NodeJS. We introduce a
bespoke error type for throwing when `Output.__str__` is called, and a
`_safe_str` helper function that catches that error type (and only that
error type). We can then use `_safe_str` in the affected logging
statements (and any others we add or identify in future) to add
best-effort logging that is safe in the presence of
`PULUMI_ERROR_OUTPUT_STRING`.

Fixes #18863