[node/components] Support resource references in schema analyzer (#18885)
This PR enables schema generation (and also the runtime, but no changes
were required for that) to infer component inputs and outputs that are
references to custom resources in Pulumi-authored providers. It doesn't
yet support third-party providers or custom resources defined ad-hoc in
the same project.

A resource reference is represented by a `$ref` like this:

```
{ $ref: "/tls/v4.11.1/schema.json#/resources/tls:index%2FprivateKey:PrivateKey" }
```

So we need to infer three parts:

- The package name which contains the referenced resource (`tls`)
- The package version (`v4.11.1`)
- The type token of the resource (`tls:index/privateKey:PrivateKey`)

This PR applies the following heuristics to do so:

- Load the type symbol for the referenced type
- Find the source file where it's declared (will point to a `d.ts` file
in `node_modules/@pulumi/foo`)
- Swap `d.ts` with `js` to find the transpiled implementation
- Require that file to load the resource class
- Lookup the `__pulumiType` static property on it to get the type token
- Use the first part of the type token to get the package name
- Find the `package.json` file in that same folder
- Read the version from it

If any of the steps fail, an error is thrown, so that we are loud and
clear about an expectation that wasn't met.

Another required piece is to fill the dependencies sections of each
language, otherwise client programs will fail to discover and load the
referenced provider. I'm not exactly sure I know how to build the
package references for each language for an arbitrary (non-Pulumi)
publisher, so I punted on this for now and hard-coded the requirement
for packages to be `@pulumi/foo`.

In terms of testing:

- I added an analyzer test that parses input and output references. For
that, I added a sample project and committed the important bits of
`node_modules`, stripping all the parts that are not required
- I extended the existing integration test to create a RandomPet
resource and return it in outputs, with both YAML and Python clients
exporting its URN and validating it's meaningful
- Tests of all numerous error cases are missing, I think I'd need to
commit too many broken `node_modules` cases to exercise those, which
doesn't seem to be a necessarily great value. It does drive code
coverage down, let me know if you think I should do more here.