Add a conformance test for `Call` provider usage
The `Call` method of the provider gRPC interface allows for the
definition and invocation of "methods" on resources, akin to method
invocations in e.g. object-oriented languages. Put another way, `Call`
is to `Invoke` as method calls are to plain function applications. In
PCL, `Call` can be expressed using the `call` primitive/intrinsic:

```hcl
x = call(resource, "methodName", args)
```

Unlike the use of `Invoke`, when a method is invoked using `Call`, there
is no need (and indeed, it is not permitted) to pass options such as
`provider` -- these are inferred from the resource on which the method
is being called (`resource` in the example above).

This change adds a conformance test that exercises these expectations --
namely, that the provider on which `Call` is, well, called, is exactly
the provider relating to the receiver resource. Initially, we add a test
for using a default provider. In a future change, we will add another
test for explicit providers, as well as tests that exercise calling
methods *on* explicit provider resources.

Part of #17025