Support calling methods on provider resources
As discussed in #18919, the `Call` method of the provider gRPC interface
allows calling methods on resources. This change extends call so that it
is possible to call methods on (explicit) *provider* resources. Doing so
requires a bit of rerouting, since the type of provider resources is
always `pulumi:providers:<package>`. Ordinarily, the built-in provider
would handle such types, but in the case of `Call`, we want the provider
receiving the call (that is, the value of the special `__self__`
argument) to *also* be the provider instance that handles the `Call`
invocation. As well as bypassing the built-in provider in such cases, we
also loosen schema binding so that `pulumi:providers:...` is permitted
as a type for functions, so that methods on providers can be defined at
the schema level.

Fixes #17025