Introduce the packages section in Pulumi.yaml (#18917)
This PR introduces a new optional section to Pulumi.yaml called
`packages`. The section can give names to package dependencies and
provide sources, versions, and arguments for those dependencies. Two
forms are supported: short and explicit:

```yaml
packages:
  planetscale:
    source: terraform-provider
    version: 0.3.0
    parameters:
    - planetscale/planetscale
  mylocalprovider: ./local/path
  agithubpackage: github.com/myorg/myrepo@v0.1.0
```

An internal type `packageValue` is defined so that we could support
deserialization of both strings and structured specs.

I've tried to keep the local folder resolution logic consistent with
what `pulumi package` does today. For that, I refactored a
`IsLocalPluginPath` helper function that is able to tell whether a given
source is a folder path or something else (remote source). If it's a
folder path, then `NewDefaultHost` will resolve those folders relative
to `ctx.Root` and add folders to the list of `projectPlugins`, similar
to what already happens for `plugins/providers`. This is currently the
only observable change in behavior.

It's the first step towards
https://github.com/pulumi/pulumi/issues/18918. Next steps:

- Update pulumi-yaml to use the new context that includes packages
- Update `pulumi install` to generate SDKs for each entry in `packages`
- Run `pulumi install` automatically for YAML runtimes?
- Update `pulumi package add` to add entries to `packages`