Run diff in parallel (#18872)
This adds a new envvar `PULUMI_PARALLEL_DIFF` which if set to true runs
step diffs in parallel. We'll roll this out as opt-in to start
(defaulting the envvar to false) but would eventually like to make this
the default because the performance wins should be significant
especially for providers with slow Diff methods.

This works by yielding from the step generator when we need to do the
diff and passing a `DiffStep` to the step executors to run in parallel.
Once the `DiffStep` runs it fulfills a `promise.CompletionSource` which
posts an event back to the step generator which is then picked up and
continues with the rest of step generation.

We're abusing Steps here to get access to the parallel but bounded step
executor workers. The DiffStep skips all the pre and post work that
normal steps would hit.

This was born out of the "run the program for refresh" workstream which
has a similar requirement for doing async work within the step
generator. I realized when discussing that with @danielrbradley that the
same idea could apply to diff.

Fixes https://github.com/pulumi/pulumi/issues/12351