typescript: import packages from @<namespace>/ (#18753)
Currently when using package add, packages are always namespaced in the
"pulumi" namespace. For typescript this means they are always named
@pulumi/<name>. However it would be better to namespace them by
namespace if that was given, so we can disambiguate them better. This PR
makes that so for TypeScript.

Part of https://github.com/pulumi/pulumi/issues/18369
Depends on https://github.com/pulumi/pulumi/pull/18418