Make protect optional (#18924)
This is so we can correctly inherit the setting from our parent, but
still allow child resources to override it. If a parent has protect set
to true, and a child explictly sets it to false then protect should not
be set on the child.

Currently because we only have true/false we treat false the same as not
set, and so a child resource can't override a parents true setting.

This has a breaking API changes. In the Go SDK the Protect field on
ConstructOptions structs are changed from a bool to a *bool. This isn't
_really_ a public API (sdk/go/common is an odd place), and a quick code
search suggests nothing is using this except us.

Note that a full fix here probably does require a breaking API change in
the Go SDK for ResourceOptions. Changing the Protect field on the
ResourceOptions struct from a bool to a *bool. The pain of that break
might be worth the clean API it results in. The alternative is to keep
Protect as bool and add a new field, something like ProtectNotSet bool.
Users wanting to check if its unset inside resource transforms, or unset
in transforms would have to look at that field. Note that this has to be
phrased as a negative because the default value for an unset bool field
in Go is false and so if we added a ProtectIsSet field that would
silently break any current transform functions where the backing runtime
code would marshal {Protect=true, ProtectIsSet=false} to protect=null.

But that decision can be left for another PR. This PR by itself still
makes everything else better.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>