Don't print "same" outputs in `--refresh` (#18943)
Operations such as `pulumi up` and `pulumi preview` support a
`--refresh` flag, which will refresh Pulumi's state before running the
subsequent operation. Under the hood, this is effectively implemented as
two operations: a refresh first, and then the requested operation, as
one might expect. Consequently, when rendering results, outputs appear
as these operations complete. For interactive use cases, users will see
`refresh` steps flashing in their list of resources; for `--diff` cases,
users will see steps appearing for both the refresh and the e.g. preview
operation.

In the latter case, the output a user sees isn't as neat as it could be
-- while much of Pulumi tries to minimise unnecessary or noisy output
(supporting flags such as `--show-sames` for users who want to see
everything), the output rendered as part of e.g. `preview --refresh
--diff` leaves a lot to be desired. For one, refreshes which have had no
effect (and are thus arguably "sames") are displayed, which can lead to
a lot of noise.

This change attempts to make the minimal tweaks required to bring
`--refresh --diff` in line with the rest of the CLI, hiding `--refresh`
steps which don't have any changes to communicate. We do this with a bit
of a hack which lets us differentiate `--refresh` steps from
(standalone) `refresh` steps by their `StepOp`, but it seems to work
well enough and the lifecycle tests that we've had to re-`PULUMI_ACCEPT`
seem to confirm it.

Fixes #18496