Make `refresh` respect `--target-dependents` (#18905)
I initially started by adding the test in this PR as a failure case.
However, as I'm already digging around in this part of the code, I added
what I think is a fix to the bug as well. This PR checks whether
`TargetDependents` is enabled when a resource is not found inside the
list of targets, and checks all its dependencies in case one of them
_is_ targeted.