Add the ability to exclude targets during updates (#18828)
We can use `--target` to list specific URNs that we'd like to operate
on, and everything else will be ignored. We can use
`--target-dependents` to say that any child of an element we `--target`
is also an implicit target. This PR implements the inverse: `--exclude`
can be used to list specific URNS to _ignore_, and everything else will
be targeted. It also adds `--exclude-dependents`, which says that any
child of an element we `--exclude` should also be excluded.

## Tests

- We can exclude a target from an update
- We can exclude a target and its dependents from an update
- We can exclude a target from a destroy
- We can exclude a target and its dependents from a destroy
- Providers of excludes are implicitly excluded from a destroy
- We can exclude a target from a refresh
- We can exclude a target and its children from a refresh

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>