Install package SDKs on pulumi install (#18940)
This change adds a new step to `pulumi install`: it will list all
packages from the `packages` section and install SDKs for each of them.
This will happen on each run of `pulumi install`, there is no check if
SDKs are already present or not, but this is supposed to be idempotent.

`package add` has no functional changes yet, it's just refactored to
expose a helper function.

I added a test with the `packages` section that runs `pulumi install`
and checks that the right thing was dowloaded. The `codecov` number
sucks but it looks like we only have integration level tests for pulumi
install and packages?

Also, a small cleanup - remove the existing folder before copying the
generated SDKs, so there is no dirty files left. Fixes
https://github.com/pulumi/pulumi-yaml/issues/755

Part of https://github.com/pulumi/pulumi/issues/18918
Fixes https://github.com/pulumi/pulumi-yaml/issues/755