Make RetainOnDelete optional on the wire protocol and DeleteBeforeReplace optional in the SDK
Following on from the fixes we did for Protect it turns out RetainOnDelete also
needed to be marked optional on the wire protocol, and DeleteBeforeReplace
needed to be made optional in the Go SDK internals (it's already optional on
the wire protocol via a separate "is defined" field).

Like with the protect fixes this doesn't change the public Go API to show that
these are optional.