Add test and fix for a DBR bug
This fixes issues where two resources both need delete/create replacing
(due to DBR diffs), but register in a different order to the origional
dependency order. This is made more likely by the parallel diff work,
but aspects of this were possible before.

We now only treat items as immpossible dependents if we've actually
processed them (not just started processing them). We now also check
after diff if another step has marked a resource as to be deleted and
restore it back to a replace.