Make protect/retain_on_delete optional
This is so we can correctly inherit the setting from our parent, but still
allow child resources to override it. If a parent has protect set to true, and
a child explictly sets it to false then protect should not be set on the child.

Currently because we only have true/false we treat false the same as not set,
and so a child resource can't override a parents true setting.

This has two breaking API changes. In the Go SDK the Protect field on the
ResourceOptions and ConstructOptions structs are changed from a bool to a
*bool. The pain of that break might be worth the clean API it results in. The
alternative is to keep Protect as bool and add a new field, something like
ProtectNotSet bool. Users wanting to check if its unset inside resource
transforms and components, or unset in transforms would have to look at that
field. Note that this has to be phrased as a negative because the default value
for an unset bool field in Go is false and so if we added a ProtectIsSet field
that would silently break any current transform functions where the backing
runtime code would marshal {Protect=true, ProtectIsSet=false} to protect=null.