nodejs: allow namespace to be explicitly set, or implicitly inferred
For components in NodeJS, we can try to infer the namespace from the
package name in the `package.json` file, if it is in the format
`@<namespace>/<name>`.

We also let the user explicitly override the namespace and name in
`PulumiPlugin.yaml`.  Those are preferred over what's in
`package.json` if given.