nodejs: infer name and namespace implicitly in run.ts
For components in NodeJS, we can try to infer the namespace from the
package name in the `package.json` file, if it is in the format
`@<namespace>/<name>`.  Do that for the automatic component provider
host.  If componentProviderHost is called explicitly make the user
pass in name and namespace manually.