python: allow namespace to be explicitly set, get rid of Metadata
In python we currently have no way to infer the namespace.  Allow it
to be set in the `PulumiPlugin.yaml` file instead.

At the same time we wanted to get rid of the Metadata class.  Make the
user specify the name in `PulumiPlugin.yaml` instead, and get rid of
that class as it is no longer necessary.