Add TODO to liftOperationType call (#18995)
I special cased this lift for pulumiResourceType/Name the other day and
wrote a comment saying we were skipping them. But found we actually
already had an issue that described how this lift was over eager and if
we fixed this generally it would cover pulumiResourceType/Name as well.
Rewording the comment and adding the link to the issue in so that we
don't have to re-discover this.