Add output verison of try generation nodejs (#18852)
For more details see: [the
doc](https://docs.google.com/document/d/1s0Ns6HppEch1_nc02zZwsVuDOR5LIMI3REXVTemhlmU)

Implement code generation for pulumi intrinsic `try` in nodejs.

A call to `try` can have parameters which result in a pulumi.Output.
This creates an issue when calling further functions on this, which
should be lowered to apply, but since try returns `any`, it does not
properly bind and call apply.

This change introduces a new program expression in codegen which is the
same as `try_` but returns an output to handle this case called
`tryOutput_` in the generated code.

To follow:
  * The same work but for `can`
  * The same work for both `try` and `can` for python

Part of #18350 and #18860

---------

Co-authored-by: Fraser Waters <fraser@pulumi.com>