infer namespace for Git Plugins; disallow overriding it in PulumiPlugin.yaml
We had a change of heart on names and namespaces.  Instead of allowing
overriding it in PulumiPlugin.yaml, we instead only want to allow
specifying them in the SDKs/inferring them from there.  For
namespaces, if the SDK doesn't infer it, we additionally want to infer
it for Git plugins. For that we use the first part of the URL after
the host.  E.g. github.com/example-org/xyz will have the namespace
`example-org`, unless otherwise specified by the component.