pass in valid info/error writers in test (#19007)
We pass in nil for the info/error writers in InstallDependencies, which
means that if there is an actual error we panic and don't have any valid
error output. Fix that by passing in `os.Stdout` and `os.Stderr`, which
go test will then print.

*Something* fails when trying to merge
https://github.com/pulumi/pulumi/actions/runs/14039014408/job/39304652576,
but it's impossible to see what because of the panic.