[components/node] Construct based on explicit component classes (#18996)
A fairly substantial change in componentProviderHost:

1. `componentProviderHost` gets a breaking change and accepts an options
interface instead of a single `string dir`. The options contain the dir
but also a map of (Component Name => Component Constructor). Only the
components passed explicitly into the map are intended to be served. The
new option is mandatory.

2. The main motivation for this change is that the `Construct`
implementation drops all dependencies on TypeScript at runtime. This
means that at execution time the plugin will no longer need to parse all
source code files on every invocation, but will simply lookup a
component constructor based on the requested name. This is probably good
for performance, but also for portability. As a follow-up step, we could
enable embedding the schema into provider artifacts, which would enable
to drop the runtime dependency on TypeScript entirely, which would
enable shipping providers compiled to JavaScript, or even to binaries.

3. The analyzer is refactored to adjust the way it discovers components.
Components are now required to be discoverable when starting with
application entry point (index.ts, or whatever is configured in
package.json). We start with the entry point, then navigate all imported
files, parsing components, until we find all components that the user
passed in. Any other components are discarded, so the schema will only
contain the ones that match the constructor map. This may make the
discovery faster, avoid parsing unrelated files and folders like tests
or examples, and keep GetSchema consistent with Construct.

We could argue that (1) makes user experience worse by requiring
explicit components, but actually this is all groundwork for
https://github.com/pulumi/pulumi/pull/18984 where we will "hide"
componentProviderHost from end users and will host exported resources
automatically. The change addresses this remark from #18984 review:

> If a user has exported a single component, but has more components non
exported, we will serve them all, which feels arbitrary. We should
probably limit to what we serve in this auto-discovery mode to the
exported ones

This approach was first implemented in the
[PoC](https://github.com/pulumi/pulumi/pull/17847) but then it didn't
get adopted in the real implementation.

I've adjusted tests for the new scenarios. You can see that `Construct`
is now easier to unit test, which is a nice side effect.