Revert change to use `t.TempDir` in testing environment (#19017)
We're seeing failures in CI on Windows when Go tries to delete the temp
dir. Revert back to the previous implementation that uses
`os.MkdirTemp`. On Windows, if there's an error deleting the temp dir,
it will be logged, but won't cause a failure. This also allows us to
keep the environment if the test fails (via `DeleteIfNotFailed`).

Fixes #19010