[components/node] Run componentProviderHost automatically in RunPlugin if any Components are exported (#18984)
This is what we have been tracking as "Automatic component inference
from code without an explicit provider implementation". This change
hides `componentProviderHost` from users, or at least makes it optional.
Instead of relying on user code to bootstrap the provider, we inspect
the exports of the loaded module. If any of the exports is a class that
extends ComponentResource, we assume that users want a component host
and call `componentProviderHost({dirname: programDir, components})`.

I updated one of the existing integration tests to illustrate the UX.