nodejs: allow namespace implicitly inferred (#18982)
For components in NodeJS, we can try to infer the namespace from the
package name in the `package.json` file, if it is in the format
`@<namespace>/<name>`.

Not sure it's worth adding an integration test here? It would be nice to
have, but it's a lot of work to set up for just this. A conformance test
would be neat, but unfortunately we don't have those running `package
add` yet.