set version to 0.0.0 for local plugins; fudge warning
For Git based components we decided to always use the Git version, and
not send back a version in the schema at all.  This version is then
used for codegen, both for example for `package.json`, and also in the
utilities file, so the engine can be told which version of the plugin
it needs.

However this doesn't work for local components, as they don't have a
"Git version". We currently try to generate the SDK with a `nil`
version, which results e.g. in TypeScript in a version field set to
${VERSION}, which makes the package not work at all.

Override the version for codegen purposes.

This also means however that the CLI will now warn users because the
version is expected to be >= 0.0.0, but we don't actually have a valid
version.  Avoid this by silencing the warning in this case.