Fix plugin resolution in convert.
Change #18951 introduced more robust plugin resolution, but convert has
some quirks that didn't set the "kind" field so resolution was not
working.  This fixes that, adds some testing, and some helpers to help
avoid this.

In the long run it'd be nice if mappers and plugin resolution used the
same functions, but this at least fixes it for now.