Fix this being incorrectly renamed in typescript when instantiating resources
When creating resources/variables when generating typescript from PCL,
the instantiation is renamed but references is when assigning attributes
specifically were overlooked.  Simply always renaming within lowering is
not acceptable since it will rewrite legitamate references to `this`.

Instead, specifically when referencing pcl attribute assigns (which
cannot reference `this` in the typescript sense, as these are only
doable when generating code as in the case of parent assign) rename
variables with valid names.

Fixes #18542