Pass list of components to component_provider_host (#19021)
Pass an explicit list of components to `component_provider_host` instead
of searching the plugin's source directory for components.

This behaviour matches that of TypeScript with
https://github.com/pulumi/pulumi/pull/18996

To find the docstrings of properties (which are not available at
runtime), we use `inspect.getsource` and parse that using the `ast`
module as before.

In a 2nd step, we can introduce a bootstrap-less mode, that detects
component plugins and runs `component_provider_host` automatically.